% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_invariance_fit.r
\name{cfa_invariance_fit}
\alias{cfa_invariance_fit}
\title{Compare fit of different CFA models}
\usage{
cfa_invariance_fit(...)
}
\arguments{
\item{...}{Four lavaan fit objects with base, weak, strong and strict invariance constraints.}
}
\description{
Compare fit of different CFA models
}
