% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_table.r
\name{cfa_table}
\alias{cfa_table}
\title{Table with items and factor loadings}
\usage{
cfa_table(object, latent_var = NULL, items = NULL, brief = TRUE,
  std = "std.all", print = FALSE, group = FALSE)
}
\arguments{
\item{object}{An object of class \code{lavaan} created by using \code{cfa} or \code{sem()}.}

\item{latent_var}{If more than one variable is specified in the CFA, which one should be shown? If left NULL, items of all latent variable will be evaluated.}

\item{items}{A data frame containing item codes and formulations.}

\item{brief}{A logical value specifying whether only standardized factor loadings should be shown.}

\item{std}{Indicates which standarized coefficient should be used (defaults to "std.all").}

\item{print}{A logical value indicating whether the values should formatted according to APA-guidelines.}

\item{new_labels}{A vector that needs to have the length of the number of effects that can be used to included a separate column with Hypotheses labels.}
}
\description{
This functions extracts the factor loadings from a lavaan object.
}
\examples{
model <- '
# latent variables
ind60 =~ x1 + x2 + x3
dem60 =~ y1 + y2 + y3 + y4
dem65 =~ y5 + y6 + y7 + y8
'
fit <- sem(model,
data = PoliticalDemocracy)

# Creating output table
cfa_table(fit, latent_var = "dem65", brief = FALSE)
}
