% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_plot.r
\name{desc_plot}
\alias{desc_plot}
\title{Plotting distributions of several likert-type variables}
\usage{
desc_plot(data, levels = 7, basic = FALSE)
}
\arguments{
\item{data}{A data.frame containing all variables that should be plotted.}

\item{levels}{Number of bins that should be plotted in the histogram. In most cases should be equal to number of answer options.}

\item{basic}{A logical value indicating whether the stylized or basis version should be plotted (defaults to FALSE = stylized version).}
}
\description{
This functions computes a facetted plot of variable distributions. Using the density function, the first layer is a histogram and the second a normal distributed density curve based on the mean and standard deviation of the respective variable.
}
