% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_vars.r
\name{describe_vars}
\alias{describe_vars}
\title{Customized version of the function describe (psych)}
\usage{
describe_vars(data, items = NULL, var_name = NULL, brief = FALSE,
  first_col = "code")
}
\arguments{
\item{data}{A data frame that includes that variables that should be evaluated. Can also be a vector with numeric values.}

\item{items}{A data frame containing the variable codes and item formulations.}

\item{var_name}{If you are evaluating a single vector, you specify the variable name here.}

\item{brief}{A logical value indicating wether all psychometrics or only the mean and the standard deviations should be printed}
}
\description{
This function is essentially similar to the function \code{describe()} in the package \code{psych} (it even uses that function), but provides some additional benefits: First, it transforms the the output into a tibble, second, it reduces number of psychometrics  to the most important ones, and it allows to include item formulation drawn from an additional data frame (needs to have a variable "code" that includes similar variable codes as the original data frame and a variable named "item" that includes the item formulations).
}
\examples{
d <- mtcars

describe_vars(mtcars)
describe_vars(mtcars$cyl, var_name = "cylinder")
}
