% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destandardize.r
\name{destandardize}
\alias{destandardize}
\title{Destandardize effect sizes (r)}
\usage{
destandardize(b = NULL, Beta = NULL, sd_x, sd_y)
}
\arguments{
\item{b}{The unstandarized effect size.}

\item{Beta}{The standardized effect size.}

\item{sd_x}{The standard deviation of the independent variable.}

\item{sd_y}{The standard deviation of the dependent variable.}
}
\description{
This function allows to destandarize (or vice versa: standardize) an effect size (Pearson's r or a typical Beta-coefficient) using the standard deviations of the independent and dependent variable.
}
\examples{
x <- rnorm(100, 2, 1)
y <- 2*x + rnorm(100, 0, 1)

destandardize(Beta = .1, sd_x = sd(x), sd_y = sd(y))
}
