% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gof.R
\name{extract_gof}
\alias{extract_gof}
\title{Computes goodness-of-fit indices from "lm" objects}
\usage{
extract_gof(..., digits = 2, print = TRUE)
}
\arguments{
\item{...}{objects of class \code{"lm"}.}
}
\description{
This function takes one of more "lm" objects and extracts goodness-of-fit indices such as AIC, BIC, logLikelihood and deviance.
}
