% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_table.r
\name{fit_table}
\alias{fit_table}
\title{Creates a table with relevant fit indices for an SEM or CFA}
\usage{
fit_table(
  object,
  indices = c("chisq", "df", "pvalue", "cfi", "tli", "rmsea", "rmsea.ci.lower",
    "rmsea.ci.upper", "srmr"),
  rmsea_ci = FALSE,
  reliability = FALSE,
  robust = FALSE,
  print = FALSE
)
}
\arguments{
\item{object}{An object of class \code{lavaan} created by using \code{cfa()} or \code{sem()} from the package 'lavaan'.}

\item{indices}{Vector of the indices that should be in the ouput.}

\item{rmsea_ci}{Logical vector indicating whether the confidence intervals of the RMSEA should be included (important for reporting the fit in an RMarkdown document using \code{\link[pmstats]{print_fit}})}

\item{reliability}{Logical value indicating whether Cronbach's alpha, McDonald's omega (composite reliability), and the average variance extracted should be included.}

\item{robust}{Logical value indicating whether the model was fitted with a robust estimator. If yes, scaled indices will be used.}

\item{print}{Logical value indicating whether the table should be formatted according to APA guidelines.}
}
\value{
A data frame containing the fit indices of the model.
}
\description{
This function returns a data frame that contains the fit indices for a SEM or CFA model. The type of indices can be chosen (available are all indices that can be obtained using the function \code{\link[lavaan]{inspect}}). Reliability indices (alpha, omega, average variance extracted) can added if desired (will be computed by using the function \code{\link[semTools]{reliability}}).
}
\examples{
library(lavaan)
model <- '
  # latent variables
  ind60 =~ x1 + x2 + x3
  dem60 =~ y1 + y2 + y3 + y4
  dem65 =~ y5 + y6 + y7 + y8

  # regressions
  dem60 ~ ind60
  dem65 ~ ind60 + dem60

  # residual covariances
  y1 ~~ y5
  y2 ~~ y4 + y6
  y3 ~~ y7
  y4 ~~ y8
  y6 ~~ y8
'
fit <- sem(model,
data = PoliticalDemocracy)

# Create table of fit indices
fit_table(fit)
fit_table(fit, indices = c("cfi", "tli"))
fit_table(fit, rmsea_ci = TRUE, print = TRUE)
}
