% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_table.r
\name{lmer_table}
\alias{lmer_table}
\title{Results table for multilevel models}
\usage{
lmer_table(object, print = FALSE)
}
\arguments{
\item{object}{An object resulting from the function \code{lmer()} from the package \code{lmerTest}.}

\item{print}{A logical value indicating whether the values should formatted according to APA-guidelines.}
}
\description{
This functions extracts the fixed effects from a lmer-model.
}
