% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_coeff.r
\name{print_coeff}
\alias{print_coeff}
\title{Prints individual coefficients for inline reporting}
\usage{
print_coeff(object, effect = NULL, variable = NULL, b = TRUE,
  se = TRUE, ci = TRUE, p = TRUE, beta = FALSE)
}
\arguments{
\item{object}{A dataframe resulting from \code{sem_table()} or \code{lmer_table()}.}

\item{effect}{A character value indicating which coefficient should be printed (draws from the label column in the dataframe).}

\item{variable}{Predictor variable that should be printed.}

\item{b}{Should the unstandardized effect be printed?}

\item{se}{Should the standard error be printed?}

\item{ci}{Should the confidence intervals be printed?}

\item{p}{Should the p-value be printed?}

\item{beta}{Should the standarized coefficient be printed?}
}
\description{
This function takes the output of \code{sem_table()} or \code{lmer_table()} and transforms it into latex-code to be used in a rmarkdown file. Individual effects in the original sem_table()-file need to be labelled. Alternatively, the predictor variable can be named. Only one coefficient at a time can be printed.
}
\examples{
model <- '
# latent variables
ind60 =~ x1 + x2 + x3
dem60 =~ y1 + y2 + y3 + y4
dem65 =~ y5 + y6 + y7 + y8

# regressions
dem60 ~ ind60
dem65 ~ ind60 + dem60

# residual covariances
y1 ~~ y5
y2 ~~ y4 + y6
y3 ~~ y7
y4 ~~ y8
y6 ~~ y8
'
fit <- sem(model,
data = PoliticalDemocracy)

# First step
results <- sem_table(fit, new_labels = c("H1", "H2", "H3))

# Second step
print_coeff(results, effect = "H1)
print_coeff(results, "H2", se = F, beta = T)
}
