% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fit.r
\name{print_fit}
\alias{print_fit}
\title{Prints fit indices for inline reporting}
\usage{
print_fit(
  fit_table,
  indices = c("chisq", "df", "pvalue", "cfi", "tli", "rmsea", "rmsea.ci.lower",
    "rmsea.ci.upper", "srmr")
)
}
\arguments{
\item{fit_table}{A dataframe resulting from \code{\link[pmstats]{fit_table}}.}

\item{indices}{Vector of the indices that should be in the ouput.}
}
\description{
This function takes the output of \code{\link[pmstats]{fit_table}} and transforms it into latex-code to be used in a rmarkdown file.
}
\examples{
model <- '
# latent variables
ind60 =~ x1 + x2 + x3
dem60 =~ y1 + y2 + y3 + y4
dem65 =~ y5 + y6 + y7 + y8

# regressions
dem60 ~ ind60
dem65 ~ ind60 + dem60

# residual covariances
y1 ~~ y5
y2 ~~ y4 + y6
y3 ~~ y7
y4 ~~ y8
y6 ~~ y8
'
fit <- sem(model,
data = PoliticalDemocracy)

# First step
table <- fit_table(fit)

# Second step
print_fit(table)
}
