% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_post_plot.r
\name{prior_post_plot}
\alias{prior_post_plot}
\title{Plotting the prior and posterior distributions of a}
\usage{
prior_post_plot(object, extra_key = FALSE, scales = "free")
}
\arguments{
\item{object}{A tibble resulting from the function \code{prior_post_values.r} or any similarly created tibble.}

\item{extra_key}{Logical value indicating whether there is more than one key to facet the plots (the second key needs to be named "key2").}

\item{scales}{Should the scales be the same of different?}
}
\value{
A ggplot object.
}
\description{
This functions takes the output of \code{prior_post_values.r} and creates overlaying density plots for each effect.
}
