% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_post_values.r
\name{prior_post_values}
\alias{prior_post_values}
\title{Extracting the prior and posterior distributions from brmsfit-objects}
\usage{
prior_post_values(object, var_names = NULL, post_effects = NULL,
  prior_means = NULL, prior_sd = NULL, n_sim = 10000)
}
\arguments{
\item{object}{An object of class \code{brmsfit}.}

\item{var_names}{Optional a vector that includes new names for the effects (variables) that should be extracted.}

\item{post_effects}{Which effect should be extracted?}

\item{prior_means}{What are means of the prior distributions (needs to be in the same order as the post_effects)?}

\item{n_sim}{How many values should be simulated for the prior distributions (at best, this number is close to the effective sample size)}
}
\value{
A tibble.
}
\description{
This functions extract the posterior distribution values from a brmsfit-object and simulates the prior distribution values
}
