% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.r
\name{render_report}
\alias{render_report}
\title{Creates an html report based on a .rmd file}
\usage{
render_report(file, name)
}
\arguments{
\item{file}{path to the file that should be rendered. File needs to be an RMarkdown script.}

\item{name}{name of the final report.}
}
\description{
This function simply allows to render RMarkdown files into html reports.
The report will be saved in the folder of the Rmarkdown file.
}
\examples{
# Produces an html-file called "01_analysis.html"
render_report("test.rmd", "01_analysis")
}
