% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_table.r
\name{sem_table}
\alias{sem_table}
\title{Result table for SEM model}
\usage{
sem_table(object, effect = NULL, labels = FALSE, regressions = FALSE,
  new_labels = NULL, one_sided = FALSE, std = "std.all",
  print = FALSE)
}
\arguments{
\item{object}{An object of class \code{lavaan} created by using \code{sem()} from the package 'lavaan'.}

\item{effect}{Vector of specific effects in the model (need to be labelled)}

\item{labels}{A logical value specifying whether only effects with labels should be included}

\item{new_labels}{A vector that needs to have the length of the number of effects that can be used to included a separate column with Hypotheses labels.}

\item{one_sided}{A logical value indicating whether the hypotheses was tested one-sided (p-values are divided by two).}

\item{print}{A logical value indicating whether the printable numbers (rounding) should be displayed.}

\item{regression}{A logical value specifying whether only regressions should be included}

\item{beta}{Indicates which standarized coefficient should be used (defaults to "std.all").}
}
\description{
This function creates a printable results table based on an SEM model. Several arguments can be specified in order to customize the output.
}
\examples{
model <- '
# latent variables
ind60 =~ x1 + x2 + x3
dem60 =~ y1 + y2 + y3 + y4
dem65 =~ y5 + y6 + y7 + y8

# regressions
dem60 ~ a*ind60
dem65 ~ b*ind60 + c*dem60

# residual covariances
y1 ~~ y5
y2 ~~ y4 + y6
y3 ~~ y7
y4 ~~ y8
y6 ~~ y8
'
fit <- sem(model,
data = PoliticalDemocracy)

# Creating output table
sem_table(fit, regressions = T, new_labels = c("H1", "H2", "H3"))
}
