% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_order_corr.r
\name{zero_order_corr}
\alias{zero_order_corr}
\title{Creates a zero-order correlation table}
\usage{
zero_order_corr(data, var_names = NULL, upper_tri = T)
}
\arguments{
\item{data}{A data frame containing all variables that should be investigated.}

\item{var_names}{A vector with fitting variable names.}

\item{upper_tri}{Should the upper triangle be omitted?}
}
\description{
Creates a classical zero-order correlation table. A vector with variable names can be added.
}
\examples{
d <- mtcars

zero_order_corr(d)
}
