% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fscores.r
\name{add_fscores}
\alias{add_fscores}
\title{Adds factor scores computed from CFA to data frame}
\usage{
add_fscores(data, object, rename_vars = NULL)
}
\arguments{
\item{data}{The data frame that passed to \code{cfa()} or \code{sem()}.}

\item{object}{An object of class \code{lavaan} created by using \code{cfa()}
or \code{sem()}.}

\item{rename_vars}{A vector including new names for the factor scores.}
}
\description{
Creates and adds factor scores to the data frame that was used to estimate the CFA.
}
\examples{
d <- PoliticalDemocracy

# Specifying model
model <- '
# latent variables
ind60 =~ x1 + x2 + x3
dem60 =~ y1 + y2 + y3 + y4
dem65 =~ y5 + y6 + y7 + y8
'
fit <- sem(model, data = d)

add_fscores(d, fit)
}
