% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_invariance_test.r
\name{cfa_invariance_test}
\alias{cfa_invariance_test}
\title{SEM model comparison (Chisquare-test)}
\usage{
cfa_invariance_test(..., model_names = NULL, print = FALSE)
}
\arguments{
\item{...}{Objects objects of class \code{lavaan} (e.g., with base, weak, strong and strict invariance constraints).}

\item{model_names}{Do you want to rename the models? Provide a vector with the respective model names.}

\item{print}{A logical value indicating whether table shoud be formatted according to APA-guidelines.}
}
\description{
This function creates a data frame that display a classical Chisqare-test of differently configured SEM/CFA models. This function is similarly written as \code{\link[pmstats]{cfa_invariance_fit}}. Both functions are meant to be used together.
}
\examples{
library(lavaan)
d <- PoliticalDemocracy \%>\% na.omit
model_1 <- '
  # latent variables
  ind60 =~ x1 + x2 + x3
  dem60 =~ y1 + y2 + y3 + y4
  dem65 =~ y5 + y6 + y7 + y8
'
fit_1 <- cfa(model_1, data = d)

model_2 <- '
  # latent variables
  ind60 =~ x1 + x2 + x3
  dem60 =~ y1 + y2 + y3
  dem65 =~ y5 + y6 + y7 + y8
'
fit_2 <- cfa(model_2, data = d)

# Model comparison
cfa_invariance_test(fit_1, fit_2)

# Renaming model names
cfa_invariance_test(fit_1, fit_2, model_names = c("Model 1: All variables", "Model 2: Some omissions"), print = TRUE)
}
