% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_invariance_test.r
\name{cfa_invariance_test}
\alias{cfa_invariance_test}
\title{Create results from invariance test}
\usage{
cfa_invariance_test(..., print = T)
}
\arguments{
\item{...}{Four lavaan fit objects with base, weak, strong and strict invariance constraints.}

\item{print}{A logical value indicating whether table shoud be formatted according to APA.}
}
\description{
This function creates a tibble that display a classical Chisq-test of differently configured models.
}
