% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_plot.r
\name{coeff_plot}
\alias{coeff_plot}
\title{Coefficient plot for a SEM}
\usage{
coeff_plot(object, effect = NULL, labels = FALSE, regressions = TRUE,
  y_limits = NULL, one_sided = FALSE)
}
\arguments{
\item{object}{An object of class \code{lavaan} created by using \code{sem()} from the package 'lavaan'.}

\item{effect}{Vector of specific effects in the model (need to be labelled in the original model)}

\item{labels}{A logical value specifying whether only effects with labels should be plotted}

\item{y_limits}{Change the limits of the x-axis.}

\item{one_sided}{A logical value indicating whether the hypotheses was tested one-sided (p-values are divided by two).}

\item{regression}{A logical value specifying whether only regressions should be plotted}
}
\description{
This function plots a classic coefficient plot for a structural equation model. More ggplot functions (e.g., theme, labs, etc.) can be passed to the resulting object.
}
\examples{
model <- '
# latent variables
ind60 =~ x1 + x2 + x3
dem60 =~ y1 + y2 + y3 + y4
dem65 =~ y5 + y6 + y7 + y8

# regressions
dem60 ~ a*ind60
dem65 ~ b*ind60 + c*dem60

# residual covariances
y1 ~~ y5
y2 ~~ y4 + y6
y3 ~~ y7
y4 ~~ y8
y6 ~~ y8
'
fit <- sem(model,
data = PoliticalDemocracy)

# Creating output table
coeff_plot(fit, regression = T, y_limits = c(0,3)) + theme_minimal()
}
