% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_table.r
\name{fit_table}
\alias{fit_table}
\title{Creates a table with relevant fit indices for an SEM of CFA}
\usage{
fit_table(object, indices = c("chisq", "df", "pvalue", "cfi", "tli",
  "rmsea", "rmsea.ci.lower", "rmsea.ci.upper", "srmr"), rmsea_ci = FALSE,
  reliability = TRUE, robust = FALSE, print = FALSE)
}
\arguments{
\item{object}{An object of class \code{lavaan} created by using \code{cfa()} or \code{sem()} from the package 'lavaan'.}

\item{indices}{Vector of the indices that should be in the ouput.}

\item{reliability}{Logical value indicating whether Cronbach's alpha, McDonald's omega (composite reliability), and the average variance extracted should be included-}

\item{robust}{Logical value indicating whether the model was fitted with a robust estimator. If yes, scaled indices will be used.}

\item{print}{Logical value indicating whether the table should be formatted according to APA guidelines.}
}
\description{
Coming soon
}
\examples{
model <- '
# latent variables
ind60 =~ x1 + x2 + x3
dem60 =~ y1 + y2 + y3 + y4
dem65 =~ y5 + y6 + y7 + y8

# regressions
dem60 ~ ind60
dem65 ~ ind60 + dem60

# residual covariances
y1 ~~ y5
y2 ~~ y4 + y6
y3 ~~ y7
y4 ~~ y8
y6 ~~ y8
'
fit <- sem(model,
data = PoliticalDemocracy)
fit_table(fit)
}
