% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderation_plot.r
\name{moderation_plot}
\alias{moderation_plot}
\title{Interaction plot for continuous moderators}
\usage{
moderation_plot(object, x1 = NULL, x2 = NULL, m = NULL,
  x_lab = "Moderator", y_lab = "Conditional effect of x on y",
  quantiles = 100, mod_hist = TRUE)
}
\arguments{
\item{object}{An object of class \code{"lm"} or \code{"lmer"} that includes an interaction.}

\item{x1}{The name of the predictor variable. Can be numeric or the first level of a factor variable.}

\item{x2}{If the factor variabel has more than two levels, this can be the second level.}

\item{m}{The name of the continuous moderator variable.}

\item{x_lab}{A character value specifying the x label in the final plot.}

\item{y_lab}{A character value specifying the y label in the final plot.}

\item{quantiles}{Number of quantiles that should be calculated.}

\item{mod_hist}{A logical value indicating whether a histogram of the moderator variable should be added above the moderation plot.}
}
\value{
If mod_hist = TRUE, this function returns a "ggExtraPlot" (i.e. a "gtable"). If mod_hist = FALSE, it returns a ggplot object that can be further customized using standard ggplot elements.
}
\description{
This functions creates a conditional effect plot that illustrates how an effect of x on y changes with changes in a continuous moderator variable. It also allows to plot a facetted version if the predictor variable x is a three-level factor.
}
\examples{
# Simulate an interaction
set.seed(10)
x <- rnorm(500, 2, 1)
z <- rnorm(500, 2, 1)
y <- 0.5*x + 1.5*(z*x) + rnorm(500, 0, 3.5)

# Estimate linear model
model <- lm(y ~ x + z + x:z)
summary(model)

# Plot model
moderation_plot(model, x1 = "x", m = "z")
}
