% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.r
\name{render_report}
\alias{render_report}
\title{Creates an html report based on a .rmd file}
\usage{
render_report(file, name)
}
\arguments{
\item{file}{path to the .rmd file that should be rendered}

\item{name}{name for the final report}
}
\description{
This function simply allows to render rmd. files into html reports. 
To be used in a "run_all.r" syntax.
}
