% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_modindices.r
\name{simple_modindices}
\alias{simple_modindices}
\title{Simple function compute modification indices}
\usage{
simple_modindices(object, arrange = TRUE, brief = TRUE, ...)
}
\arguments{
\item{object}{An object of class lavaan.}

\item{arrange}{Should the modification indices be sorted from large to small?}

\item{brief}{If true, only modification indices are shown.}

\item{...}{Further argument that can be passed to the function \code{modindices()}.}
}
\description{
This is simply a convenient function that allows to extract modification indices from a fitted lavaan object. It uses the function \code{\link[lavaan]{modindices}} to compute modification indices, yet provides additional options to arrange them in descending order and exclude any additional information.
}
