% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_reduction.r
\name{variance_reduction}
\alias{variance_reduction}
\title{Variance reduction for two-level random intercept models}
\usage{
variance_reduction(base_model, ..., red = TRUE)
}
\arguments{
\item{base_model}{baseline model with no predictors}

\item{...}{up to four comparison models with predictors}

\item{red}{a logical value indicating wether only the explained variance should be reported (defaults to TRUE).}
}
\description{
Function to compute the amount explained variance on each level by a random-intercept multilevel model. 
The amount of explained variance is computed by comparing the relevant models to the baseline (null-) model (i.e., the bare random-intercept model without any predictors).
}
