% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_reduction.r
\name{variance_reduction}
\alias{variance_reduction}
\title{Explained variance for two-level random intercept models}
\usage{
variance_reduction(object, ..., digits = 2, print = FALSE)
}
\arguments{
\item{object}{An object of class merMod (more specifically, an object of subclass lmerMod). This first object needs to be the null model with no predictors. All other models will be compared to this baseline model.}

\item{...}{Up to four objects of class merMod.}

\item{red}{A logical value indicating wether only the explained variance should be reported (defaults to TRUE).}
}
\value{
A data fram containing the explained variance on both levels.
}
\description{
Function to compute the amount explained variance on each level by a random-intercept multilevel model. 
The amount of explained variance is computed by comparing the relevant models to the baseline (null-) model (i.e., the bare random-intercept model without any predictors).
}
\examples{
m0 <- lme4::lmer(Reaction ~ 1 + (1 | Subject), sleepstudy)
m1 <- lme4::lmer(Reaction ~ 1 + Days + (1 | Subject), sleepstudy)

variance_reduction(m0, m1, print = T, digits = 3)
}
